server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["pybash1 <example@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
fastapi-login = "^1.8.2"
python-dotenv = "^0.20.0"
SQLAlchemy = "^1.4.39"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
